Use dynamically allocated arrays for codebooks
This is in preparation for allowing the huffman coding to represent any
output. In zlib we need to decode values in the range [0,285]. This can
currently not be represented in our codebook. Use a dynamic array for
the code book, and store the number of inputs that were used to
construct the huffman tree on generation. At this point this should have
no effect since the only generator is using a 256 length array of u8
character frequencies