Replace huffman tree walking with array lookup
With canonical huffman coding, all elements for a given bit width end up
with sequential codes. This allows for an optimization where we can
convert a huffman code to the appropriate value by just looking up the
parsed value in an array.

Store an array of values for each bit width. When reading, start by
reading the minimum number of bits and try reading the corresponding
table. If the value is too large, read another bit and try the next one.

This results in an ~2x speedup in our parsing, some in the bit reader, a
lot in the actual value lookup.

The code for generating the arrays looks very similar for the code to
generate the code book. Pull out an iterator that generates codes and
use it for both

Generating the codebook no longer requires the tree structure, so remove
it from the huffman table
  * Note that the implementation of codebook generation no longer needs
    to store the actual tree, just the bit depths. We can do the same
    logic for building the tree, but discard the branches immediately
    and just increment the node depth on each iteration

There's probably more in here, review before merging was minimal